requirements.txt
tensorflow==2.13.0
tensorflow-hub==0.14.0
streamlit==1.28.1
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
